Refactor server into modular structure
